Add namespaced builtin filters and sources
Add new style namespaced filters and sources.
This commit does not break compatibilities (users still can use old styles).

    New style                   Old style

    file/old                    file_old
    file/point                  file_point
    file/rec                    file_rec

    converter/abbr_word         converter_abbr_word
    converter/relative_abbr     converter_relative_abbr
    converter/relative_word     converter_relative_word

... etc
