+++ rplugin/python3/denite/filter/converter_abbr_word.py
@@ -1,22 +1,8 @@
-# ============================================================================
-# FILE: converter_abbr_word.py
-# AUTHOR: Shougo Matsushita <Shougo.Matsu at gmail.com>
-# License: MIT license
-# ============================================================================
-
-from .base import Base
+# For backward compatibility
+from .converter.abbr_word import Filter as Base
 
 
 class Filter(Base):
-
     def __init__(self, vim):
         super().__init__(vim)
-
-        self.name = 'converter_abbr_word'
-        self.description = 'convert candidate abbr to word'
-
-    def filter(self, context):
-        for candidate in context['candidates']:
-            if 'abbr' in candidate:
-                candidate['word'] = candidate['abbr']
-        return context['candidates']
+        self.name = self.name.replace('/', '_')
