+++ rplugin/python3/denite/filter/matcher_hide_hidden_files.py
@@ -1,25 +1,8 @@
-# ============================================================================
-# FILE: matcher_hide_hidden_files.py
-# AUTHOR: Shougo Matsushita <Shougo.Matsu at gmail.com>
-# License: MIT license
-# ============================================================================
-
-from .base import Base
-from os.path import basename
-from re import match
+# For backward compatibility
+from .matcher.hide_hidden_files import Filter as Base
 
 
 class Filter(Base):
-
     def __init__(self, vim):
         super().__init__(vim)
-
-        self.name = 'matcher_hide_hidden_files'
-        self.description = 'hide the hidden files'
-
-    def filter(self, context):
-        if '.' in context['input']:
-            return context['candidates']
-
-        return [x for x in context['candidates']
-                if not match('\.', basename(x['action__path']))]
+        self.name = self.name.replace('/', '_')
